HW-2 repository 이미지 검색 적용
