+++ JayYi/ArchitectureStudy/app/src/main/java/com/jay/architecturestudy/ui/image/ImageFragment.kt
@@ -5,6 +5,7 @@ import android.util.Log
 import com.jay.architecturestudy.R
 import com.jay.architecturestudy.data.model.Image
 import com.jay.architecturestudy.data.model.ResponseNaverQuery
+import com.jay.architecturestudy.data.repository.NaverSearchRepositoryImpl
 import com.jay.architecturestudy.network.Api
 import com.jay.architecturestudy.ui.BaseFragment
 import com.jay.architecturestudy.util.toPx
@@ -17,6 +18,10 @@ import retrofit2.Response
 class ImageFragment : BaseFragment(R.layout.fragemnt_image) {
 
     private lateinit var imageAdapter: ImageAdapter
+
+    private val naverSearchRepository by lazy {
+        NaverSearchRepositoryImpl()
+    }
 
     override fun onActivityCreated(savedInstanceState: Bundle?) {
         super.onActivityCreated(savedInstanceState)
@@ -34,22 +39,14 @@ class ImageFragment : BaseFragment(R.layout.fragemnt_image) {
     }
 
     override fun search(keyword: String) {
-        Api.getImages(keyword)
-            .enqueue(object : Callback<ResponseNaverQuery<Image>> {
-                override fun onFailure(call: Call<ResponseNaverQuery<Image>>, t: Throwable) {
-                    Log.e("Image", "error=${t.message}")
-                }
-
-                override fun onResponse(
-                    call: Call<ResponseNaverQuery<Image>>,
-                    response: Response<ResponseNaverQuery<Image>>
-                ) {
-                    if (response.isSuccessful) {
-                        val body = response.body() ?: return
-                        imageAdapter.setData(body.items)
-                    }
-                }
-
-            })
+        naverSearchRepository.getImage(
+            keyword = keyword,
+            success = {responseImage ->
+                imageAdapter.setData(responseImage.images)
+            },
+            fail = { e ->
+                Log.e("Image", "error=${e.message}")
+            }
+        )
     }
 }
